+++ integration/src/main/java/com/quartercode/quarterbukkit/QuarterBukkitIntegration.java
@@ -127,7 +127,7 @@ public class QuarterBukkitIntegration {
         Bukkit.getLogger().info("Querying server mods api ...");
 
         // Get latest version
-        List<ProjectFile> avaiableFiles = new FilesQuery(PROJECT_ID, new VersionParser() {
+        List<ProjectFile> availableFiles = new FilesQuery(PROJECT_ID, new VersionParser() {
 
             @Override
             public String parseVersion(ProjectFile file) {
@@ -135,11 +135,11 @@ public class QuarterBukkitIntegration {
                 return file.getName().replace("QuarterBukkit ", "");
             }
         }).execute();
-        if (avaiableFiles.size() == 0) {
-            // No file avaiable
+        if (availableFiles.size() == 0) {
+            // No file available
             return;
         }
-        ProjectFile latestFile = avaiableFiles.get(avaiableFiles.size() - 1);
+        ProjectFile latestFile = availableFiles.get(availableFiles.size() - 1);
 
         Bukkit.getLogger().info("Found the latest version of " + PLUGIN_NAME + ": " + latestFile.getVersion());