+++ plugin/src/main/java/com/quartercode/quarterbukkit/util/QuarterBukkitUpdater.java
@@ -60,7 +60,7 @@ public class QuarterBukkitUpdater {
             plugin.getLogger().info("Querying server mods api ...");
 
             // Get latest version
-            List<ProjectFile> avaiableFiles = new FilesQuery(PROJECT_ID, new VersionParser() {
+            List<ProjectFile> availableFiles = new FilesQuery(PROJECT_ID, new VersionParser() {
 
                 @Override
                 public String parseVersion(ProjectFile file) {
@@ -68,12 +68,12 @@ public class QuarterBukkitUpdater {
                     return file.getName().replace("QuarterBukkit ", "");
                 }
             }).execute();
-            if (avaiableFiles.size() == 0) {
-                // No file avaiable
+            if (availableFiles.size() == 0) {
+                // No file available
                 return false;
             }
 
-            ProjectFile latestFile = avaiableFiles.get(avaiableFiles.size() - 1);
+            ProjectFile latestFile = availableFiles.get(availableFiles.size() - 1);
             if (latestFile.getVersion().equals(plugin.getDescription().getVersion())) {
                 // No update required (latest version already installed)
                 return false;